consolidate marker webgl functions
